tests/wpt/tests/html/canvas/offscreen/text/2d.text.draw.fill.unaffected.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.draw.fill.unaffected</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.draw.fill.unaffected</h1>
<p class="desc">fillText does not start a new path or subpath</p>


<script>
var t = async_test("fillText does not start a new path or subpath");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.moveTo(0, 0);
  ctx.lineTo(100, 0);

  ctx.font = '35px Arial, sans-serif';
  ctx.fillText('FAIL', 5, 35);

  ctx.lineTo(100, 50);
  ctx.lineTo(0, 50);
  ctx.fillStyle = '#0f0';
  ctx.fill();

  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 5,45, 0,255,0,255);
  t.done();

});
</script>
